mini-job-search/requirements.txt
flask
psycopg2-binary
requests
openai
